+++ app/src/main/java/com/jcrawleydev/gemsdrop/view/BorderLayer.java
@@ -38,13 +38,13 @@ public class BorderLayer implements DrawableItem {
 
 
     @Override
-    public int getX() {
-        return 0;
+    public float getX() {
+        return 0f;
     }
 
 
     @Override
-    public int getY() {
+    public float getY() {
         return 0;
     }
